// apps/exam-vault/src/generated/models/response-model.ts
import type { Attempt } from './attempt-model';
import type { Question } from './question-model';

export const ResponseGradingStatusKeyToLabel = {
  'GradingStatusKey0': 'Pending',
  'GradingStatusKey1': 'AutoGraded',
  'GradingStatusKey2': 'ManuallyGraded'
} as const;
export type ResponseGradingStatusKey = keyof typeof ResponseGradingStatusKeyToLabel;

export interface Response {
  /**
   * @displayName Response ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Response Summary
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  responseSummary: string;
  /**
   * @displayName Answer Text
   * @validationRule Required for create/update operations
   */
  answerText: string;
  /**
   * @displayName Attempt
   * @validationRule Required for create/update operations
   */
  attempt: Pick<Attempt, 'id' | 'studentName'>;
  /**
   * @displayName Audio URL
   */
  audioURL?: string;
  /**
   * @displayName Grader Feedback
   */
  graderFeedback?: string;
  /**
   * @displayName Grading Status
   * @validationRule Required for create/update operations
   */
  gradingStatusKey: ResponseGradingStatusKey;
  /**
   * @displayName Is Correct
   */
  isCorrect?: boolean;
  /**
   * @displayName Points Earned
   */
  pointsEarned?: number;
  /**
   * @displayName Question
   * @validationRule Required for create/update operations
   */
  question: Pick<Question, 'id' | 'questionTitle'>;
}

export const _Response = 'Response' as const;